+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.33.0] - 2020-10-26
 
+### Added
+- DatePicker: Added tests for Input, onEnterPress event, test for pressing the "Backspace", "Esc" and "Tab"
+
 ### Fixed
 - DropDownSelect: Removed dropdown opening logic from focus handler to click handler 
 - MultiSelect: Fixed checkbox classes